requirements/development.txt
-r base.txt

django-extensions
ipython
werkzeug
